Moving everything to separate files
