+++ EXP-Client/App.js
@@ -1,14 +1,12 @@
-import { StatusBar } from 'expo-status-bar';
 import * as React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { MD3LightTheme, Provider } from 'react-native-paper';
 import Constants from 'expo-constants';
-import axios from 'axios';
 
-import TaskItem from './TaskItem'
-import ArchivedDay from './ArchivedDay'
-import Todo from './Todo'
-import Nav from './Nav'
+import TaskItem from './todo/TaskList';
+import ArchivedDay from './progress/ArchivedDay';
+import Todo from './todo/Todo';
+import Nav from './components/Nav';
 
 function testingCode() {
   return (
